test: move fake service into testing assets
